+++ kube-derive/tests/crd_schema_test.rs
@@ -9,6 +9,7 @@ use serde::{Deserialize, Serialize};
     group = "clux.dev",
     version = "v1",
     kind = "Foo",
+    category = "clux",
     namespaced,
     derive = "PartialEq"
 )]
@@ -56,6 +57,7 @@ fn test_crd_schema_matches_expected() {
             "spec": {
                 "group": "clux.dev",
                 "names": {
+                    "categories": ["clux"],
                     "kind": "Foo",
                     "plural": "foos",
                     "shortNames": [],